Move E2EE and emote module to their own directories
